build a debian package
This uses cargo-deb to generate a `.deb` package that can be installed
using `apt install -f ./haxo.deb`. This command will also install the
required dependencies (alsa and fluidsynth).

This will only work on bullseye for now: bookworm uses a newer version
of fluidsynth so we would need to build a separate deb.