cross compile using zigbuild
Uses cargo-zigbuild to cross compile for RPi targets. Rather than let
pkgconfig find the libraries in system locations, we download and unpack
the required RPi packages into the libs/ subfolder, the configure
cargo-zigbuild to find them.

This hardcodes the package versions, so will only work on Bullseye and
might break if packages are updated. A better approach might be to run
`apt download` on the target device and SCP the `.deb` files back to the
build machine.

Includes Makefiles to make it easy to download all required libraries and
run the build command.